// src/typechain/DescribeSignature.ts
export type SubSignatureStructOutput = {
  kind: number;
  signer: string;
  signature: string;
};

export type AccountDescriptionStruct = {
  kind: PromiseOrValue<BigNumberish>;
  has7702Delegation: PromiseOrValue<boolean>;
  owners: PromiseOrValue<string>[];
  threshold: PromiseOrValue<BigNumberish>;
};

export type AccountDescriptionStructOutput = {
  kind: number;
  has7702Delegation: boolean;
  owners: string[];
  threshold: BigNumber;
};

export type SignatureDataStruct = {
  kind: PromiseOrValue<BigNumberish>;
  signer: PromiseOrValue<string>;
  subSignatures: SubSignatureStruct[];
  account: AccountDescriptionStruct;
};

export type SignatureDataStructOutput = {
  kind: number;
  signer: string;
  subSignatures: SubSignatureStructOutput[];
  account: AccountDescriptionStructOutput;
};

export interface DescribeSignatureInterface extends utils.Interface {
  functions: {
    "describeSignature(address,bytes,bytes)": FunctionFragment;
    "encodeMessageDataForSafe(address,bytes)": FunctionFragment;
    "getMessageHashForSafe(address,bytes)": FunctionFragment;
  };

  getFunction(
    nameOrSignatureOrTopic:
      | "describeSignature"
      | "encodeMessageDataForSafe"
      | "getMessageHashForSafe"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "describeSignature",
    values: [
      PromiseOrValue<string>,
      PromiseOrValue<BytesLike>,
      PromiseOrValue<BytesLike>
    ]
  ): string;
  encodeFunctionData(
    functionFragment: "encodeMessageDataForSafe",
    values: [PromiseOrValue<string>, PromiseOrValue<BytesLike>]
  ): string;
  encodeFunctionData(
    functionFragment: "getMessageHashForSafe",
    values: [PromiseOrValue<string>, PromiseOrValue<BytesLike>]
  ): string;

  decodeFunctionResult(
    functionFragment: "describeSignature",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "encodeMessageDataForSafe",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getMessageHashForSafe",
    data: BytesLike
  ): Result;

  events: {};
}

export interface DescribeSignature extends BaseContract {
  contractName: "DescribeSignature";

  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: DescribeSignatureInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    describeSignature(
      signer: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      signature: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<
      [SignatureDataStructOutput] & { data: SignatureDataStructOutput }
    >;

    encodeMessageDataForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<[string]>;

    getMessageHashForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<[string]>;
  };

  describeSignature(
    signer: PromiseOrValue<string>,
    message: PromiseOrValue<BytesLike>,
    signature: PromiseOrValue<BytesLike>,
    overrides?: CallOverrides
  ): Promise<SignatureDataStructOutput>;

  encodeMessageDataForSafe(
    safe: PromiseOrValue<string>,
    message: PromiseOrValue<BytesLike>,
    overrides?: CallOverrides
  ): Promise<string>;

  getMessageHashForSafe(
    safe: PromiseOrValue<string>,
    message: PromiseOrValue<BytesLike>,
    overrides?: CallOverrides
  ): Promise<string>;

  callStatic: {
    describeSignature(
      signer: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      signature: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<SignatureDataStructOutput>;

    encodeMessageDataForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<string>;

    getMessageHashForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<string>;
  };

  filters: {};

  estimateGas: {
    describeSignature(
      signer: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      signature: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    encodeMessageDataForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    getMessageHashForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    describeSignature(
      signer: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      signature: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    encodeMessageDataForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    getMessageHashForSafe(
      safe: PromiseOrValue<string>,
      message: PromiseOrValue<BytesLike>,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}